+++ src/Window.hpp
@@ -23,7 +23,7 @@ class Window{
         void ToggleFullscreen();
         void Draw(sf::Drawable& l_drawable);
         void Zoom(float l_zoom);
-        sf::Vector2i convertToGameCoordinates(sf::Vector2i viewportCoordinates);
+        sf::Vector2f convertToGameCoordinates(sf::Vector2i viewportCoordinates);
         sf::RenderWindow m_window;
         sf::Vector2f getScaleOfWindowSizeToView();
         private:
